+++ packages/gatsby-remark-images/src/index.js
@@ -4,10 +4,12 @@ const {
   imageWrapperClass,
 } = require(`./constants`)
 const visitWithParents = require(`unist-util-visit-parents`)
+const getDefinitions = require(`mdast-util-definitions`)
 const path = require(`path`)
+const queryString = require(`query-string`)
 const isRelativeUrl = require(`is-relative-url`)
 const _ = require(`lodash`)
-const { fluid } = require(`gatsby-plugin-sharp`)
+const { fluid, traceSVG } = require(`gatsby-plugin-sharp`)
 const Promise = require(`bluebird`)
 const cheerio = require(`cheerio`)
 const slash = require(`slash`)
@@ -30,6 +32,7 @@ module.exports = (
     showCaptions: false,
     pathPrefix,
     withWebp: false,
+    tracedSVG: false,
   }
 
   const options = _.defaults(pluginOptions, defaults)
@@ -40,6 +43,9 @@ module.exports = (
         (node.type === `html` && !!node.value.match(/<a /)) ||
         node.type === `link`
     )
+
+  // Get all the available definitions in the markdown tree
+  const definitions = getDefinitions(markdownAST)
 
   // This will allow the use of html image tags
   // const rawHtmlNodes = select(markdownAST, `html`)
@@ -53,21 +59,42 @@ module.exports = (
   // This will only work for markdown syntax image tags
   let markdownImageNodes = []
 
-  visitWithParents(markdownAST, `image`, (node, ancestors) => {
-    const inLink = ancestors.some(findParentLinks)
+  visitWithParents(
+    markdownAST,
+    [`image`, `imageReference`],
+    (node, ancestors) => {
+      const inLink = ancestors.some(findParentLinks)
 
-    markdownImageNodes.push({ node, inLink })
-  })
+      markdownImageNodes.push({ node, inLink })
+    }
+  )
+
+  const getImageInfo = uri => {
+    const { url, query } = queryString.parseUrl(uri)
+    return {
+      ext: path
+        .extname(url)
+        .split(`.`)
+        .pop(),
+      url,
+      query,
+    }
+  }
 
   // Takes a node and generates the needed images and then returns
   // the needed HTML replacement for the image
-  const generateImagesAndUpdateNode = async function(node, resolve, inLink) {
+  const generateImagesAndUpdateNode = async function(
+    node,
+    resolve,
+    inLink,
+    overWrites = {}
+  ) {
     // Check if this markdownNode has a File parent. This plugin
     // won't work if the image isn't hosted locally.
     const parentNode = getNode(markdownNode.parent)
     let imagePath
     if (parentNode && parentNode.dir) {
-      imagePath = slash(path.join(parentNode.dir, node.url))
+      imagePath = slash(path.join(parentNode.dir, getImageInfo(node.url).url))
     } else {
       return null
     }
@@ -100,10 +127,18 @@ module.exports = (
     const presentationWidth = fluidResult.presentationWidth
 
     // Generate default alt tag
-    const srcSplit = node.url.split(`/`)
+    const srcSplit = getImageInfo(node.url).url.split(`/`)
     const fileName = srcSplit[srcSplit.length - 1]
     const fileNameNoExt = fileName.replace(/\.[^/.]+$/, ``)
     const defaultAlt = fileNameNoExt.replace(/[^A-Z0-9]/gi, ` `)
+
+    const alt = overWrites.alt
+      ? overWrites.alt
+      : node.alt
+      ? node.alt
+      : defaultAlt
+
+    const title = node.title ? node.title : ``
 
     const imageStyle = `
       width: 100%;
@@ -123,8 +158,8 @@ module.exports = (
       <img
         class="${imageClass}"
         style="${imageStyle}"
-        alt="${node.alt ? node.alt : defaultAlt}"
-        title="${node.title ? node.title : ``}"
+        alt="${alt}"
+        title="${title}"
         src="${fallbackSrc}"
         srcset="${srcSet}"
         sizes="${fluidResult.sizes}"
@@ -165,11 +200,37 @@ module.exports = (
           class="${imageClass}"
           style="${imageStyle}"
           src="${fallbackSrc}"
-          alt="${node.alt ? node.alt : defaultAlt}"
-          title="${node.title ? node.title : ``}"
+          alt="${alt}"
+          title="${title}"
         />
       </picture>
       `.trim()
+    }
+
+    let placeholderImageData = fluidResult.base64
+
+    // if options.tracedSVG is enabled generate the traced SVG and use that as the placeholder image
+    if (options.tracedSVG) {
+      let args = typeof options.tracedSVG === `object` ? options.tracedSVG : {}
+
+      // Translate Potrace constants (e.g. TURNPOLICY_LEFT, COLOR_AUTO) to the values Potrace expects
+      const { Potrace } = require(`potrace`)
+      const argsKeys = Object.keys(args)
+      args = argsKeys.reduce((result, key) => {
+        const value = args[key]
+        result[key] = Potrace.hasOwnProperty(value) ? Potrace[value] : value
+        return result
+      }, {})
+
+      const tracedSVG = await traceSVG({
+        file: imageNode,
+        args,
+        fileArgs: args,
+        reporter,
+      })
+
+      // Escape single quotes so the SVG data can be used in inline style attribute with single quotes
+      placeholderImageData = tracedSVG.replace(/'/g, `\\'`)
     }
 
     const ratio = `${(1 / fluidResult.aspectRatio) * 100}%`
@@ -185,9 +246,7 @@ module.exports = (
   >
     <span
       class="${imageBackgroundClass}"
-      style="padding-bottom: ${ratio}; position: relative; bottom: 0; left: 0; background-image: url('${
-      fluidResult.base64
-    }'); background-size: cover; display: block;"
+      style="padding-bottom: ${ratio}; position: relative; bottom: 0; left: 0; background-image: url('${placeholderImageData}'); background-size: cover; display: block;"
     ></span>
     ${imageTag}
   </span>
@@ -226,7 +285,24 @@ module.exports = (
     markdownImageNodes.map(
       ({ node, inLink }) =>
         new Promise(async (resolve, reject) => {
-          const fileType = node.url.slice(-3)
+          const overWrites = {}
+          let refNode
+          if (
+            !node.hasOwnProperty(`url`) &&
+            node.hasOwnProperty(`identifier`)
+          ) {
+            //consider as imageReference node
+            refNode = node
+            node = definitions(refNode.identifier)
+            // pass original alt from referencing node
+            overWrites.alt = refNode.alt
+            if (!node) {
+              // no definition found for image reference,
+              // so there's nothing for us to do.
+              return resolve()
+            }
+          }
+          const fileType = getImageInfo(node.url).ext
 
           // Ignore gifs as we can't process them,
           // svgs as they are already responsive by definition
@@ -238,11 +314,15 @@ module.exports = (
             const rawHTML = await generateImagesAndUpdateNode(
               node,
               resolve,
-              inLink
+              inLink,
+              overWrites
             )
 
             if (rawHTML) {
-              // Replace the image node with an inline HTML node.
+              // Replace the image or ref node with an inline HTML node.
+              if (refNode) {
+                node = refNode
+              }
               node.type = `html`
               node.value = rawHTML
             }
@@ -286,7 +366,7 @@ module.exports = (
                 return resolve()
               }
 
-              const fileType = formattedImgTag.url.slice(-3)
+              const fileType = getImageInfo(formattedImgTag.url).ext
 
               // Ignore gifs as we can't process them,
               // svgs as they are already responsive by definition