+++ packages/gatsby-source-filesystem/src/create-remote-file-node.js
@@ -6,10 +6,19 @@ const { isWebUri } = require(`valid-url`)
 const Queue = require(`better-queue`)
 const readChunk = require(`read-chunk`)
 const fileType = require(`file-type`)
+const ProgressBar = require(`progress`)
 
 const { createFileNode } = require(`./create-file-node`)
 const { getRemoteFileExtension, getRemoteFileName } = require(`./utils`)
 const cacheId = url => `create-remote-file-node-${url}`
+
+const bar = new ProgressBar(
+  `Downloading remote files [:bar] :current/:total :elapsed secs :percent`,
+  {
+    total: 0,
+    width: 30,
+  }
+)
 
 /********************
  * Type Definitions *
@@ -114,7 +123,8 @@ async function pushToQueue(task, cb) {
     const node = await processRemoteNode(task)
     return cb(null, node)
   } catch (e) {
-    return cb(null, e)
+    console.warn(`Failed to process remote content ${task.url}`)
+    return cb(e)
   }
 }
 
@@ -135,7 +145,7 @@ async function pushToQueue(task, cb) {
 const requestRemoteNode = (url, headers, tmpFilename) =>
   new Promise((resolve, reject) => {
     const responseStream = got.stream(url, {
-      ...headers,
+      headers,
       timeout: 30000,
       retries: 5,
     })
@@ -173,9 +183,11 @@ async function processRemoteNode({
   store,
   cache,
   createNode,
+  parentNodeId,
   auth = {},
   createNodeId,
   ext,
+  name,
 }) {
   // Ensure our cache directory exists.
   const pluginCacheDir = path.join(
@@ -202,7 +214,9 @@ async function processRemoteNode({
 
   // Create the temp and permanent file names for the url.
   const digest = createHash(url)
-  const name = getRemoteFileName(url)
+  if (!name) {
+    name = getRemoteFileName(url)
+  }
   if (!ext) {
     ext = getRemoteFileExtension(url)
   }
@@ -210,47 +224,42 @@ async function processRemoteNode({
   const tmpFilename = createFilePath(pluginCacheDir, `tmp-${digest}`, ext)
 
   // Fetch the file.
-  try {
-    const response = await requestRemoteNode(url, headers, tmpFilename)
+  const response = await requestRemoteNode(url, headers, tmpFilename)
+
+  if (response.statusCode == 200) {
     // Save the response headers for future requests.
     await cache.set(cacheId(url), response.headers)
-
-    // If the user did not provide an extension and we couldn't get one from remote file, try and guess one
-    if (ext === ``) {
-      const buffer = readChunk.sync(tmpFilename, 0, fileType.minimumBytes)
-      const filetype = fileType(buffer)
-      if (filetype) {
-        ext = `.${filetype.ext}`
-      }
-    }
-
-    const filename = createFilePath(
-      path.join(pluginCacheDir, digest),
-      name,
-      ext
-    )
-    // If the status code is 200, move the piped temp file to the real name.
-    if (response.statusCode === 200) {
-      await fs.move(tmpFilename, filename, { overwrite: true })
-      // Else if 304, remove the empty response.
-    } else {
-      await fs.remove(tmpFilename)
-    }
-
-    // Create the file node.
-    const fileNode = await createFileNode(filename, createNodeId, {})
-    fileNode.internal.description = `File "${url}"`
-    // Override the default plugin as gatsby-source-filesystem needs to
-    // be the owner of File nodes or there'll be conflicts if any other
-    // File nodes are created through normal usages of
-    // gatsby-source-filesystem.
-    createNode(fileNode, { name: `gatsby-source-filesystem` })
-
-    return fileNode
-  } catch (err) {
-    // ignore
   }
-  return null
+
+  // If the user did not provide an extension and we couldn't get one from remote file, try and guess one
+  if (ext === ``) {
+    const buffer = readChunk.sync(tmpFilename, 0, fileType.minimumBytes)
+    const filetype = fileType(buffer)
+    if (filetype) {
+      ext = `.${filetype.ext}`
+    }
+  }
+
+  const filename = createFilePath(path.join(pluginCacheDir, digest), name, ext)
+  // If the status code is 200, move the piped temp file to the real name.
+  if (response.statusCode === 200) {
+    await fs.move(tmpFilename, filename, { overwrite: true })
+    // Else if 304, remove the empty response.
+  } else {
+    await fs.remove(tmpFilename)
+  }
+
+  // Create the file node.
+  const fileNode = await createFileNode(filename, createNodeId, {})
+  fileNode.internal.description = `File "${url}"`
+  fileNode.parent = parentNodeId
+  // Override the default plugin as gatsby-source-filesystem needs to
+  // be the owner of File nodes or there'll be conflicts if any other
+  // File nodes are created through normal usages of
+  // gatsby-source-filesystem.
+  createNode(fileNode, { name: `gatsby-source-filesystem` })
+
+  return fileNode
 }
 
 /**
@@ -278,6 +287,9 @@ const pushTask = task =>
       })
   })
 
+// Keep track of the total number of jobs we push in the queue
+let totalJobs = 0
+
 /***************
  * Entry Point *
  ***************/
@@ -298,9 +310,11 @@ module.exports = ({
   store,
   cache,
   createNode,
+  parentNodeId = null,
   auth = {},
   createNodeId,
   ext = null,
+  name = null,
 }) => {
   // validation of the input
   // without this it's notoriously easy to pass in the wrong `createNodeId`
@@ -332,13 +346,23 @@ module.exports = ({
     return Promise.resolve()
   }
 
-  return (processingCache[url] = pushTask({
+  totalJobs += 1
+  bar.total = totalJobs
+
+  const fileDownloadPromise = pushTask({
     url,
     store,
     cache,
     createNode,
+    parentNodeId,
     createNodeId,
     auth,
     ext,
-  }))
+    name,
+  })
+
+  fileDownloadPromise.then(() => bar.tick())
+
+  processingCache[url] = fileDownloadPromise
+  return fileDownloadPromise
 }