+++ packages/gatsby/src/bootstrap/__tests__/resolve-module-exports.js
@@ -111,6 +111,10 @@ describe(`Resolve module exports`, () => {
       exports.__esModule = true;
       exports.foo = '';
     `,
+    "/export/named": `const foo = ''; export { foo };`,
+    "/export/named/from": `export { Component } from 'react';`,
+    "/export/named/as": `const foo = ''; export { foo as bar };`,
+    "/export/named/multiple": `const foo = ''; const bar = ''; const baz = ''; export { foo, bar, baz };`,
   }
 
   beforeEach(() => {
@@ -173,5 +177,25 @@ describe(`Resolve module exports`, () => {
   it(`Ignores exports.__esModule`, () => {
     const result = resolveModuleExports(`/esmodule/export`, resolver)
     expect(result).toEqual([`foo`])
+  })
+
+  it(`Resolves a named export`, () => {
+    const result = resolveModuleExports(`/export/named`, resolver)
+    expect(result).toEqual([`foo`])
+  })
+
+  it(`Resolves a named export from`, () => {
+    const result = resolveModuleExports(`/export/named/from`, resolver)
+    expect(result).toEqual([`Component`])
+  })
+
+  it(`Resolves a named export as`, () => {
+    const result = resolveModuleExports(`/export/named/as`, resolver)
+    expect(result).toEqual([`bar`])
+  })
+
+  it(`Resolves multiple named exports`, () => {
+    const result = resolveModuleExports(`/export/named/multiple`, resolver)
+    expect(result).toEqual([`foo`, `bar`, `baz`])
   })
 })