+++ www/src/components/navigation.js
@@ -149,6 +149,7 @@ const Navigation = ({ pathname }) => {
             <NavItem linkTo="/features/">Features</NavItem>
             <NavItem linkTo="/blog/">Blog</NavItem>
             <NavItem linkTo="/showcase/">Showcase</NavItem>
+            <NavItem linkTo="/contributing/">Contributing</NavItem>
             {/* <li css={styles.li}>
                 <Link
                   to="/community/"
@@ -180,16 +181,10 @@ const Navigation = ({ pathname }) => {
               [presets.Hd]: { display: `flex` },
             }}
           >
-            <SocialNavItem
-              href="https://discord.gg/0ZcbPKXt5bVoxkfV"
-              title="Discord"
-            >
+            <SocialNavItem href="https://gatsby.app/discord" title="Discord">
               <DiscordIcon overrideCSS={{ verticalAlign: `text-top` }} />
             </SocialNavItem>
-            <SocialNavItem
-              href="https://twitter.com/gatsbyjs"
-              title="@gatsbyjs"
-            >
+            <SocialNavItem href="https://twitter.com/gatsbyjs" title="Twitter">
               <TwitterIcon style={{ verticalAlign: `text-top` }} />
             </SocialNavItem>
           </div>