+++ packages/gatsby-remark-images/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/compare/gatsby-remark-images@3.0.4...gatsby-remark-images@3.0.5) (2019-02-22)
+
+### Features
+
+- **gatsby-remark-images:** Add option to use tracedSVG instead of blur up effect ([#9490](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/issues/9490)) ([072bcdd](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/commit/072bcdd))
+
+## [3.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/compare/gatsby-remark-images@3.0.3...gatsby-remark-images@3.0.4) (2019-02-15)
+
+### Bug Fixes
+
+- **gatsby-remark-images:** ensure query string is ignored when detecting images ([#11743](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/issues/11743)) ([37cc21f](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/commit/37cc21f))
+
+## [3.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/compare/gatsby-remark-images@3.0.2...gatsby-remark-images@3.0.3) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-remark-images
+
+## [3.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/compare/gatsby-remark-images@3.0.1...gatsby-remark-images@3.0.2) (2019-01-31)
+
+### Features
+
+- **gatsby-remark-images:** handle image references ([#11086](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/issues/11086)) ([9ee6b4a](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/commit/9ee6b4a))
+
 <a name="3.0.1"></a>
 
 ## [3.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-images/compare/gatsby-remark-images@3.0.0...gatsby-remark-images@3.0.1) (2018-11-29)